Refactor AssetProofComposer (#333)
