+++ ledger/src/integration-test/java/com/scalar/dl/ledger/service/LedgerServiceEndToEndTest.java
@@ -8,8 +8,6 @@ import static com.scalar.dl.ledger.service.Constants.ASSET_ID_COLUMN_NAME;
 import static com.scalar.dl.ledger.service.Constants.ASSET_OUTPUT_COLUMN_NAME;
 import static com.scalar.dl.ledger.service.Constants.AUDITOR_ENTITY_ID;
 import static com.scalar.dl.ledger.service.Constants.BALANCE_ATTRIBUTE_NAME;
-import static com.scalar.dl.ledger.service.Constants.CERTIFICATE_A;
-import static com.scalar.dl.ledger.service.Constants.CERTIFICATE_B;
 import static com.scalar.dl.ledger.service.Constants.CONTRACT_ID_ATTRIBUTE_NAME;
 import static com.scalar.dl.ledger.service.Constants.CREATE_CONTRACT_ID1;
 import static com.scalar.dl.ledger.service.Constants.CREATE_CONTRACT_ID2;
@@ -38,8 +36,6 @@ import static com.scalar.dl.ledger.service.Constants.PAYMENT_CONTRACT_ID1;
 import static com.scalar.dl.ledger.service.Constants.PAYMENT_CONTRACT_ID2;
 import static com.scalar.dl.ledger.service.Constants.PAYMENT_CONTRACT_ID3;
 import static com.scalar.dl.ledger.service.Constants.PAYMENT_CONTRACT_ID4;
-import static com.scalar.dl.ledger.service.Constants.PRIVATE_KEY_A;
-import static com.scalar.dl.ledger.service.Constants.PRIVATE_KEY_B;
 import static com.scalar.dl.ledger.service.Constants.SECRET_KEY_A;
 import static com.scalar.dl.ledger.service.Constants.SECRET_KEY_B;
 import static com.scalar.dl.ledger.service.Constants.SOME_AMOUNT_1;
@@ -49,6 +45,10 @@ import static com.scalar.dl.ledger.service.Constants.SOME_ASSET_ID_1;
 import static com.scalar.dl.ledger.service.Constants.SOME_ASSET_ID_2;
 import static com.scalar.dl.ledger.service.Constants.SOME_CIPHER_KEY;
 import static com.scalar.dl.ledger.service.Constants.SOME_ID;
+import static com.scalar.dl.ledger.test.TestConstants.CERTIFICATE_A;
+import static com.scalar.dl.ledger.test.TestConstants.CERTIFICATE_B;
+import static com.scalar.dl.ledger.test.TestConstants.PRIVATE_KEY_A;
+import static com.scalar.dl.ledger.test.TestConstants.PRIVATE_KEY_B;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.catchThrowable;
