+++ ledger/src/integration-test/java/com/scalar/dl/ledger/service/LedgerServiceIntegrationTest.java
@@ -4,8 +4,6 @@ import static com.scalar.dl.ledger.service.Constants.AMOUNT_ATTRIBUTE_NAME;
 import static com.scalar.dl.ledger.service.Constants.ASSETS_ATTRIBUTE_NAME;
 import static com.scalar.dl.ledger.service.Constants.ASSET_ATTRIBUTE_NAME;
 import static com.scalar.dl.ledger.service.Constants.BALANCE_ATTRIBUTE_NAME;
-import static com.scalar.dl.ledger.service.Constants.CERTIFICATE_A;
-import static com.scalar.dl.ledger.service.Constants.CERTIFICATE_B;
 import static com.scalar.dl.ledger.service.Constants.CONTRACT_ID_ATTRIBUTE_NAME;
 import static com.scalar.dl.ledger.service.Constants.CREATE_CONTRACT_ID1;
 import static com.scalar.dl.ledger.service.Constants.CREATE_CONTRACT_ID2;
@@ -23,8 +21,6 @@ import static com.scalar.dl.ledger.service.Constants.GET_BALANCE_CONTRACT_ID4;
 import static com.scalar.dl.ledger.service.Constants.ID_ATTRIBUTE_NAME;
 import static com.scalar.dl.ledger.service.Constants.KEY_VERSION;
 import static com.scalar.dl.ledger.service.Constants.PAYMENT_CONTRACT_ID1;
-import static com.scalar.dl.ledger.service.Constants.PRIVATE_KEY_A;
-import static com.scalar.dl.ledger.service.Constants.PRIVATE_KEY_B;
 import static com.scalar.dl.ledger.service.Constants.SECRET_KEY_A;
 import static com.scalar.dl.ledger.service.Constants.SECRET_KEY_B;
 import static com.scalar.dl.ledger.service.Constants.SOME_AMOUNT_1;
@@ -35,6 +31,10 @@ import static com.scalar.dl.ledger.service.Constants.SOME_ASSET_ID_2;
 import static com.scalar.dl.ledger.service.Constants.SOME_BALANCE;
 import static com.scalar.dl.ledger.service.Constants.SOME_ID;
 import static com.scalar.dl.ledger.service.Constants.SOME_NONCE;
+import static com.scalar.dl.ledger.test.TestConstants.CERTIFICATE_A;
+import static com.scalar.dl.ledger.test.TestConstants.CERTIFICATE_B;
+import static com.scalar.dl.ledger.test.TestConstants.PRIVATE_KEY_A;
+import static com.scalar.dl.ledger.test.TestConstants.PRIVATE_KEY_B;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
