+++ ledger/src/main/java/com/scalar/dl/ledger/service/LedgerValidationService.java
@@ -7,9 +7,9 @@ import com.scalar.dl.ledger.contract.ContractManager;
 import com.scalar.dl.ledger.crypto.ClientKeyValidator;
 import com.scalar.dl.ledger.crypto.SignatureValidator;
 import com.scalar.dl.ledger.database.AssetFilter;
+import com.scalar.dl.ledger.database.AssetProofComposer;
 import com.scalar.dl.ledger.database.Transaction;
 import com.scalar.dl.ledger.database.TransactionManager;
-import com.scalar.dl.ledger.database.scalardb.AssetProofComposer;
 import com.scalar.dl.ledger.error.LedgerError;
 import com.scalar.dl.ledger.exception.LedgerException;
 import com.scalar.dl.ledger.exception.ValidationException;
