+++ ledger/src/main/java/com/scalar/dl/ledger/service/ValidationService.java
@@ -5,8 +5,8 @@ import com.scalar.dl.ledger.contract.ContractEntry;
 import com.scalar.dl.ledger.contract.ContractMachine;
 import com.scalar.dl.ledger.contract.ContractManager;
 import com.scalar.dl.ledger.crypto.ClientKeyValidator;
+import com.scalar.dl.ledger.database.AssetProofComposer;
 import com.scalar.dl.ledger.database.TransactionManager;
-import com.scalar.dl.ledger.database.scalardb.AssetProofComposer;
 import com.scalar.dl.ledger.model.LedgerValidationRequest;
 import com.scalar.dl.ledger.model.LedgerValidationResult;
 import com.scalar.dl.ledger.statemachine.DeserializationType;
